@@ -1,4 +1,4 @@
 [tool.black]
 line-length = 200
-target-version = ['py311']
+target-version = ['py312']
 include = '\.pyi?$'
