@@ -1,5 +1,6 @@
-from disnake import ApplicationCommandInteraction
-from disnake.ext import commands
+import discord
+from discord import ApplicationContext
+from discord.ext import commands
 
 from client import Client
 
@@ -7,19 +8,17 @@ from client import Client
 class Github(commands.Cog):
     """A set of commands for interacting with the GitHub API."""
 
-    @commands.slash_command()
-    async def github(self, inter: ApplicationCommandInteraction) -> None:
-        del inter
+    github = discord.SlashCommandGroup("github", "Interact with GitHub.")
 
-    @github.sub_command()
-    async def user(self, inter: ApplicationCommandInteraction) -> None:
+    @github.command()
+    async def user(self, context: ApplicationContext) -> None:
         """Retrieves information about a GitHub user."""
-        await inter.send("This command is a work in progress.")
+        await context.respond("This command is a work in progress.")
 
-    @github.sub_command()
-    async def repository(self, inter: ApplicationCommandInteraction) -> None:
+    @github.command()
+    async def repository(self, context: ApplicationContext) -> None:
         """Retrieves information about a GitHub repository."""
-        await inter.send("This command is a work in progress.")
+        await context.respond("This command is a work in progress.")
 
 
 def setup(client: Client):
