@@ -1,4 +1,4 @@
 import subprocess
 
-version = "0.1.0"
+version = "0.0.1"
 hash = subprocess.check_output(["git", "rev-parse", "--short", "HEAD"]).strip().decode("utf8")
